# [aarch64] speed up increment and decrement operators with overflow de…