@sebpop 2019-05-06
The patch reduces the total number of instructions executed as counted by callgrind:
```
ZEND_DONT_UNLOAD_MODULES=1 valgrind --tool=callgrind --separate-recs=1 --dump-instr=yes --cache-sim=no sapi/cgi/php-cgi -d opcache.jit=0 -T100 wordpress/index.php > /dev/null
```
before the patch: `I refs: 2,524,336,840`
after the patch: `I refs: 2,523,322,562`
That is `0.04%` fewer instructions.