@nikic 2019-05-07
Has been merged with https://github.com/php/php-src/commit/79a84e00fad27bab2a4ca7a58ace1f68cc88be5c into 7.4.